app/requirements.txt
fastapi==0.111.0
langchain==0.2.0
langchain_community==0.2.0
langchain_core==0.2.0
langchain_experimental==0.0.59
langchain_openai==0.1.7
langserve==0.2.0
python-dotenv==1.0.1
Requests==2.31.0
sse_starlette==1.6.1
